Refactor sender.ts to move filtering code into separate module
