+++ x-pack/plugins/security_solution/server/lib/telemetry/sender.ts
@@ -21,27 +21,15 @@ import { TelemetryDiagTask } from './diagnostic_task';
 import { TelemetryEndpointTask } from './endpoint_task';
 import { EndpointAppContextService } from '../../endpoint/endpoint_app_context_services';
 import { AgentService, AgentPolicyServiceInterface } from '../../../../fleet/server';
+import {
+  TelemetryEvent
+} from './types';
+import {
+  AllowlistAlertEventFields,
+  copyAllowlistedFields
+} from './filters'
 
-type BaseSearchTypes = string | number | boolean | object;
-export type SearchTypes = BaseSearchTypes | BaseSearchTypes[] | undefined;
 
-export interface TelemetryEvent {
-  [key: string]: SearchTypes;
-  '@timestamp'?: string;
-  data_stream?: {
-    [key: string]: SearchTypes;
-    dataset?: string;
-  };
-  cluster_name?: string;
-  cluster_uuid?: string;
-  file?: {
-    [key: string]: SearchTypes;
-    Ext?: {
-      [key: string]: SearchTypes;
-    };
-  };
-  license?: ESLicense;
-}
 
 export class TelemetryEventsSender {
   private readonly initialCheckDelayMs = 10 * 1000;
@@ -276,7 +264,7 @@ export class TelemetryEventsSender {
 
   public processEvents(events: TelemetryEvent[]): TelemetryEvent[] {
     return events.map(function (obj: TelemetryEvent): TelemetryEvent {
-      return copyAllowlistedFields(allowlistEventFields, obj);
+      return copyAllowlistedFields(AllowlistAlertEventFields, obj);
     });
   }
 
@@ -437,147 +425,7 @@ export class TelemetryEventsSender {
   }
 }
 
-// For the Allowlist definition.
-interface AllowlistFields {
-  [key: string]: boolean | AllowlistFields;
-}
 
-// Allow list process fields within events.  This includes "process" and "Target.process".'
-const allowlistProcessFields: AllowlistFields = {
-  args: true,
-  name: true,
-  executable: true,
-  command_line: true,
-  hash: true,
-  pid: true,
-  pe: {
-    original_file_name: true,
-  },
-  uptime: true,
-  Ext: {
-    architecture: true,
-    code_signature: true,
-    dll: true,
-    malware_signature: true,
-    token: {
-      integrity_level_name: true,
-    },
-  },
-  thread: true,
-  working_directory: true,
-};
-
-// Allow list for event-related fields, which can also be nested under events[]
-const allowlistBaseEventFields: AllowlistFields = {
-  dll: {
-    name: true,
-    path: true,
-    code_signature: true,
-    malware_signature: true,
-    pe: {
-      original_file_name: true,
-    },
-  },
-  event: true,
-  file: {
-    extension: true,
-    name: true,
-    path: true,
-    size: true,
-    created: true,
-    accessed: true,
-    mtime: true,
-    directory: true,
-    hash: true,
-    Ext: {
-      code_signature: true,
-      header_data: true,
-      malware_classification: true,
-      malware_signature: true,
-      quarantine_result: true,
-      quarantine_message: true,
-    },
-  },
-  process: {
-    parent: allowlistProcessFields,
-    ...allowlistProcessFields,
-  },
-  network: {
-    direction: true,
-  },
-  registry: {
-    data: {
-      strings: true,
-    },
-    hive: true,
-    key: true,
-    path: true,
-    value: true,
-  },
-  Target: {
-    process: {
-      parent: allowlistProcessFields,
-      ...allowlistProcessFields,
-    },
-  },
-  user: {
-    id: true,
-  },
-};
-
-// Allow list for the data we include in the events. True means that it is deep-cloned
-// blindly. Object contents means that we only copy the fields that appear explicitly in
-// the sub-object.
-const allowlistEventFields: AllowlistFields = {
-  '@timestamp': true,
-  agent: true,
-  Endpoint: true,
-  /* eslint-disable @typescript-eslint/naming-convention */
-  Memory_protection: true,
-  Ransomware: true,
-  data_stream: true,
-  ecs: true,
-  elastic: true,
-  // behavioral protection re-nests some field sets under events.*
-  events: allowlistBaseEventFields,
-  rule: {
-    id: true,
-    name: true,
-    ruleset: true,
-    version: true,
-  },
-  host: {
-    os: true,
-  },
-  ...allowlistBaseEventFields,
-};
-
-export function copyAllowlistedFields(
-  allowlist: AllowlistFields,
-  event: TelemetryEvent
-): TelemetryEvent {
-  return Object.entries(allowlist).reduce<TelemetryEvent>((newEvent, [allowKey, allowValue]) => {
-    const eventValue = event[allowKey];
-    if (eventValue !== null && eventValue !== undefined) {
-      if (allowValue === true) {
-        return { ...newEvent, [allowKey]: eventValue };
-      } else if (typeof allowValue === 'object' && Array.isArray(eventValue)) {
-        const subValues = eventValue.filter((v) => typeof v === 'object');
-        return {
-          ...newEvent,
-          [allowKey]: subValues.map((v) => copyAllowlistedFields(allowValue, v as TelemetryEvent)),
-        };
-      } else if (typeof allowValue === 'object' && typeof eventValue === 'object') {
-        const values = copyAllowlistedFields(allowValue, eventValue as TelemetryEvent);
-        return {
-          ...newEvent,
-          ...(Object.keys(values).length > 0 ? { [allowKey]: values } : {}),
-        };
-      }
-    }
-    return newEvent;
-  }, {});
-}
 
 // Forms URLs like:
 // https://telemetry.elastic.co/v3/send/my-channel-name or
